[PATCH] Update driver for Chaos 4.4 and fix for HW error status
The kernel on which Chaos 4.4 is removed pci_find_aer_capability
function and replaces it with a call to pci_find_ext_capability.

The HCA checks for certain hardware errors which can be falsely
triggered when the link is reset. The fix is to mask them rather
than report them.

Signed-off-by: Mitko Haralanov <mitko@qlogic.com>