[PATCH] if qib_init() fails, driver fails to clean up properly
If qib_init() fails, the driver fails to free memory, unregister device
files, and unregister with the PCIe framework. The driver will unload
without error but a subsequent driver load will cause the system to panic.
This was found by changing the 7220 code to load the serdes microcode
separately and not installing the microcode file.

Signed-off-by: Ralph Campbell <ralph.campbell@qlogic.com>