[PATCH] fix race between qib_error_qp() and receive packet processing
When transitioning a QP to the error state, in progress RWQEs need
to be marked complete. This also involves releasing the reference
count to the memory regions referenced in the SGEs. The locking in
the receive packet processing wasn't sufficient to prevent
qib_error_qp() from modifying the r_sge state at the same time,
thus leading to kernel panics.

Signed-off-by: Ralph Campbell <ralph.campbell@qlogic.com>