IB/qib: change rcv queue/qpn selection
The basic idea is that on SusieQ, the difficult part of mapping QPN to context
is handled by the mapping registers so the generic QPN allocation doesn't need
to worry about chip specifics. For Monty and Linda, there is no mapping table
so the qpt->mask (same as dd->qpn_mask), is used to see if the QPN to context
falls within [zero..dd->n_krcv_queues).

Signed-off-by: Mike Marciniszyn <mike.marciniszyn@qlogic.com>