IB/qib: Issue pre-emptive NAKs on eager buffer overflow
Under congestion resulting in eager buffer overflow attempt to send
pre-emptive NAKs if header queue entries with TID errors are generated and
a valid header is present. This prevents long timeouts and flow restarts if
a trailing set of packets are dropped due to eager overflows. Pre-emptive
NAKs are currently only supported for RDMA Writes.

Signed-off-by: Avneesh Pant <avneesh.pant@qlogic.com>