IB/qib: Un-necessary delayed completions on RC connection
Currently on receipt of a response message (ACKs, RDMA Response, Atomic
Responses etc.) if the SDMA completion counter is not advanced the
driver delays the completion of the WQE. In most cases this is overly
pessimistic as the response (ACK) to a previously transmitted send
implies that the send is complete. Ensure that SDMA queue is progressed
appropriately before determining if a send has delayed completions.

Signed-off-by: Avneesh Pant <avneesh.pant@qlogic.com>