IB/qib: Fix potential deadlock with link down interrupt
There is a possibility of a deadlock due to the way locks are acquired and
released in qib_set_uevent_bits(). The function qib_set_uevent_bits() is called
in process context calling spin_lock() and spin_unlock() while this same lock
is acquired/released in interrupt context which can lead to a deadlock when
running on the same cpu.

The fix is to replace spin_lock() and spin_unlock() with spin_lock_irqsave()
and spin_unlock_irqrestore() respectively in qib_set_uevent_bits().

Signed-off-by: Ram Vepa <ram.vepa@qlogic.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@qlogic.com>