IB/qib: Ensure that LOS and DFE are being turned off
Due to timing, it is possible for the LOS and DFE to remain
on. This is due to the link progressing to LinkUP prior to
the driver getting the first Status Changed interrupt.
By expanding the conditions under which LOS is turned off
and DFE timeout is being set, timing is no longer an issue.

Signed-off-by: Mike Marciniszyn <mike.marciniszyn@qlogic.com>