IB/qib: memcpy optimizations
The default memcpy used by qib_copy_sge() ends up being a rep movsb on x86_64,
which is pretty slow.

This fix adds an x86_64 specific routine that 1) probes for X86_FEATURE_REP_GOOD
and 2) uses an inline asm routine builton rep movsq that testing has shown is
better than the builtin memcpy for all cases up to 4K.  The probing routine is
now called when the qib module is loaded to enable the optimization.   When
X86_FEATURE_REP_GOOD is not set, the routine uses the kernel's unrolled __memcpy
when the length is more than 64 and the builtin memcpy otherwise.

The existing copy as well as a new stub probe routine are maintained as weak
symbols for other architectures.

This patch also adds the cache bypass copies from older releases.

The cache_bypass_copy module parameter is used to enable vs
default copy_sge code.

The qib_verbs_send_dma() and qib_copy_from_sge are also changed to use the
memcpy_string_op().

Signed-off-by: Mike Marciniszyn <mike.marciniszyn@qlogic.com>