IB/qib: remove s_lock around header validation
Observation in qib_ruc_check_hdr() shows that the s_lock is not required
in the normal case.   The r_lock is held in all cases, and protects
the qp fields that are read.

The s_lock will be needed to around the call to qib_migrate_qp() to
insure that the send engine sees a consistent set of fields.

Signed-off-by: Mike Marciniszyn <mike.marciniszyn@qlogic.com>