IB/qib: Adjust to use cpumask_weight() and cpumask_first()
The cpus_weight macro doesn't work correctly when the
CONFIG_CPUMASK_OFFSTACK is chosen.

The macro can walk run off the end of the valid bits in the
mask because NR_CPUS is much larger than the actual number of bits.

This patch adjusts the code to use the APIs that were added
for kernels >= 2.6.30.  Backport adjustments will need to be
made for older kernels.

An equivalent patch was also submitted externally by
KOSAKI Motohiro <kosaki.motohiro@jp.fujitsu.com>.

Signed-off-by: Mike Marciniszyn <mike.marciniszyn@qlogic.com>