IB/qib: Optimize locking for get_txreq()
The current code locks the QP s_lock, followed by the pending_lock, I guess
to to protect against the allocate failing.

This patch only locks the pending_lock, assuming that the empty case is an
exeception, in which case the pending_lock is dropped, and the original code is
executed.  This will save a lock of s_lock in the normal case.

The observation is that the sdma descriptors will deplete at twice the rate of
txreq's, so this should be rare.

Signed-off-by: Mike Marciniszyn <mike.marciniszyn@qlogic.com>