IB/qib: add module paramater for sdma_idle_cnt
Testing with some platforms has shown that if this value is
increased, then bidirectional performance tests improve when
the test is running on a foreign NUMA node.

Increasing the value increases the time the chip will wait
before signaling an idle interrupt, giving the processor more
time to supply packets to the transmitter and reducing the
number of sdma idle interrupts.

The patch doesn't change the existing default of 64.

Signed-off-by: Mike Marciniszyn <mike.marciniszyn@qlogic.com>