IB/qib: Add logic for affinity hint
Call irq_set_affinity_hint to provide user mode programs
like irqbalance the information to be able to distribute
qib interrupts appropriately.

The logic allocates all non-receive interrupts to the first
CPU local to the HCA.  Receive interrupts are allocated round
robin starting with the second CPU local to the HCA with
potential wrap back to the second CPU.

In the special case where the hca's local mask is the same
weight as the number of cpus, cpus local to CPU0 are used.

Signed-off-by: Mike Marciniszyn <mike.marciniszyn@qlogic.com>