IB/qib: Improve MPI message rates on NUMA platforms
Allow a user application to optimize the reads of the send buffer status and
the available receive entry registers with memory IO or memory operations or
both based on the proximity of the user process from the numa node on which
receive and send status memory is allocated. A single memory location for all
sends and receives are shared among all processes. This change allows the
process to optimize the reads dynamically such that the processes running on
the numa node on which the allocated memory resides opt for memory operations
while the other processes choose the memory IO operations. This change
significantly improves the MPI message rates on the latest NUMA platforms.
    
Signed-off-by: Ramkrishna Vepa <ram.vepa@qlogic.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@qlogic.com>