IB/qib: optimize pio ack buffer allocation
This patch optimizes pio buffer allocation in the kernel.

For the qib, kernel pio buffers are used for sending acks.  The code
to allocate the buffer would always start at 0 until it found a buffer.

This means that an average of 64 comparisions were done on each allocate, since
the busy bit won't be cleared until the bits are refreshed when buffers are
exhausted.

This patch adds two new fields in the devdata struct, last_pio and
min_kernel_pio.  last_pio is the last buffer that was allocated.
min_kernel_pio is the lowest potential available buffer.

min_kernel_pio is modifed as contexts are allocated and
deallocted.

Reviewed-by: Ram Vepa <ram.vepa@qlogic.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@qlogic.com>