IB/qib: Fix qib driver mkey lease timeout handling
If a port has a MKey lease set, the MKey protect bits set to 1,
and a SubnSet arrives with an invalid MKey, a MKey mismatch trap is
generated, the lease timer begins as expected, and eventually the MKey
protect bits will be set back to 0 as per the spec.  However, if any
other SMP with an invalid MKey arrives, the lease timer is expired and
the MKey protect bits remain in force.

This is not according to to spec.  In particular, C14-17 says that
a lease timer that is underway is not affected by protection level
checks (ie, at protection level 1, a SubnGet with a bad MKey may be
successful, but does not stop the timer), and C14-19 says that the timer
shall stop when a valid MKey has been received.  C14-19 is the only
compliance statement that specifies a stopping condition for the timer.

This behavior is magnified if the port's Master SM LID attribute
points at itself.  In that case, the MKey mismatch trap is sufficient to
expire the timer, and the mkey lease attribute is rendered useless.

Reviewed-by: Ram Vepa <ram.vepa@qlogic.com>
Signed-off-by: Jim Foraker <foraker1@llnl.gov>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@qlogic.com>