IB/qib: patch for ordering issue with 16 bytes writes at tail of move
This patch corrects an issue where a 16 byte move at the tail of a bypass move
will store the last 8 bytes prior to the preceding 8 bytes.

OpenMPI must do some memory polling on memory being updated with an RDMA write
and the sequencing causes issues with the subsequent processing.

The flaw exists in both memcpy_cachebypass() and memcpy_cachebypass2().

Reviewed-by: Mike Heinz <michael.william.heinz@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>