IB/qib: correct ordering of reregister vs. port active event
When a port first goes active with SMA Set(PortInfo) and reregister bit set, the
driver sends up the reregister event followed by a port active event.

The problem is that in response to reregister event most apps try to issue a SA
query of some sort, but that fails because port is not active.

The qib driver needs to a trivial change to correct this behavior.

This issue has been there for a while, however the recent serdes
work has probably made the delay between the reregister event and the active
event larger and hence opened the race far enough so that its being seen more
often.

This fragment in subn_set_portinfo needs to move:

if (pip->clientrereg_resv_subnetto & 0x80) {
        clientrereg = 1;
        event.event = IB_EVENT_CLIENT_REREGISTER;
        ib_dispatch_event(&event);
}

This patch moves the code to just before the nested call to subn_get_portinfo():

    TO HERE
    ret = subn_get_portinfo(smp, ibdev, port);

The patch also changes the clientrereg local to a u8 and saves off the rereg bit into
it.  The code following the nested subn_get_portinfo() now restores that bit per o14-12.2.1
with a logical from that copy.

Reviewed-by: Ram Vepa <ram.vepa@qlogic.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@qlogic.com>