IB/qib: Avoid returning EBUSY from MR deregister
A timing issue can occur where qib_mr_dereg can
return -EBUSY if the MR use count has is not zero.

This can occur if the MR is deregistered while RDMA
read response packets are being progressed from the
SDMA ring.  The suspicion is that the peer sent an RDMA
read request, which has already been copied across to the
peer.  The peer sees the completion of his request and
then communicates to the responder that the MR is not
needed any longer.  The responder tries to deregister the
MR, catching some responses remaining in the SDMA ring holding
the MR use count.

The code now uses a get/put paradigm to track MR use
counts and coordinates with the MR de-registration
process using a completion when the count has reached
zero.  A timeout on the delay is in place to catch
other EBUSY issues.

The reference count protocol is as follows:
- The return to the user counts as 1
- A reference from the lk_table or the qib_ibdev counts as 1.
- Transient I/O operations increase/decrease as necessary

Additional cleanup:
- Push dma_mr key allocation into qib_keys.c
- Simply interface to alloc/dealloc lkey routines

Reviewed-by: Ram Vepa <ram.vepa@qlogic.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>