IB/qib: Fix message rate degradation on AMD Magny Cours & Interlagos
The optimal message rate performance with PSM on AMD Magny Cours and
Interlagos systems is acheived when the receive header queue, eager
buffers and the send pioavail register are allocated on the local
NUMA node. The driver allocates these memories on the NUMA node on
which the process is running. The fix is to allow for both these
memory allocation schemas.
If the numa_aware loadable parameter is not set which would be the
case for the AMD systems the driver allocates these memories on
the HCA's local NUMA node which can be overriden by a loadable
parameter, numa_node. A configuration of an invalid NUMA node
causes the driver to allocate memories on the local NUMA node.
If the numa_aware loadable parameter is set by the user or is
set automatically by the driver then the receive buffers are
allocated on the numa node on which the process is running and
the send pioavail register is allocated on the local NUMA node.

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Ramkrishna Vepa <ramkrishna.vepa@intel.com>