IB/qib: Add dual rail NUMA awareness for PSM processes
The driver currently selects a HCA based on the algorithm that
PSM chooses, contexts within a HCA or across. The HCA can also
be chosen by the user. Either way, this patch assigns a CPU on
the NUMA node local to the selected HCA. This patch also tries
to select the HCA closest to the NUMA node of the CPU assigned
via taskset to PSM process. If this HCA is unusable then
another unit is selected based on the algorithm that is
currently enforced or selected by PSM - round robin
context selection 'within' or 'across' HCA's.

Fixed a bug wherein contexts are setup on the NUMA node on
which the processes are opened (setup_ctxt()) and not on the
NUMA node that the driver recommends the CPU on.

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Ramkrishna Vepa <ramkrishna.vepa@intel.com>