IB/qib: Correct CPU0 sibling fallback for affinity hint
When the affinity hint logic detected that the number of local cpus
was invalid it tried to determine the siblings of CPU0 to use
as the cpumask.

Unfortunately the wrong function was used to get the mask, which
provided the full mask of online CPUs.  This was incorrect for CPUs
that interleave the CPUs between sockets.

The patch corrects the issue.

Signed-off-by: Mike Marciniszyn <mike.marciniszyn@qlogic.com>